Starting the Google Books Api fetch

+++ src/main/java/org/example/splitbooks/entity/Genre.java
@@ -10,7 +10,7 @@ public class Genre {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int genreId;
+    private Long genreId;
 
     @Column(unique = true, nullable = false)
     private String genreName;
